move attribute_base classes outside of model
