+++ lib/csv_row_model/concerns/dynamic_columns_base.rb
@@ -1,5 +1,4 @@
 require 'csv_row_model/concerns/model/dynamic_columns'
-require "byebug"
 
 # Shared between Import and Export, see test fixture for basic setup
 module CsvRowModel
