+++ spec/csv_row_model/internal/dynamic_column_attribute_base_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe CsvRowModel::Model::DynamicColumnAttribute do
+describe CsvRowModel::DynamicColumnAttributeBase do
   describe "instance" do
     let(:instance) { described_class.new(:skills, row_model) }
     let(:row_model_class) do
